authentication and authorisation using passport-jwt

+++ controllers/api/v1/posts_api.js
@@ -22,16 +22,23 @@ module.exports.index = async function (req, res) {
 module.exports.destroy = async function(req, res) {
     try {
         let post =  await Post.findById(req.params.id);
-        await Post.findByIdAndDelete(req.params.id);
-        await Comment.deleteMany({post: req.params.id});
-        console.log('Succesfully deleted post along with comments');
 
-        return res.status(200).json({
-            message: "Post and associated comments deleted"
-        });
+        if (post.user == req.user.id) {
+            await Post.findByIdAndDelete(req.params.id);
+            await Comment.deleteMany({post: req.params.id});
+            console.log('Succesfully deleted post along with comments');
+    
+            return res.json(200, {
+                message: "Post and associated comments deleted"
+            });
+        } else {
+            return res.json(401, {
+                message: "Unauthorized user, you cannot delete this post"
+            })
+        }
     } catch (err) {
         console.log('Error in deleting post and comments', err);
-        return res.status(500).json({
+        return res.json(500, {
             message: "Internal Sever Error!"
         });
     }
